Move files to util.file
