+++ maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/action/ConvertAction.java
@@ -15,7 +15,7 @@ import org.vincentyeh.IMG2PDF.commandline.parser.core.HandledException;
 import org.vincentyeh.IMG2PDF.converter.PDFConverter;
 import org.vincentyeh.IMG2PDF.task.Task;
 import org.vincentyeh.IMG2PDF.task.TaskList;
-import org.vincentyeh.IMG2PDF.util.FileChecker;
+import org.vincentyeh.IMG2PDF.util.file.FileChecker;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
