+++ maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/util/file/FileChecker.java
@@ -1,4 +1,4 @@
-package org.vincentyeh.IMG2PDF.util;
+package org.vincentyeh.IMG2PDF.util.file;
 
 import org.vincentyeh.IMG2PDF.SharedSpace;
