+++ maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/util/file/FileFilterHelper.java
@@ -1,4 +1,4 @@
-package org.vincentyeh.IMG2PDF.util;
+package org.vincentyeh.IMG2PDF.util.file;
 
 import java.io.File;
 import java.io.FileFilter;
